% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_report.R
\name{cache_report}
\alias{cache_report}
\title{Cache report}
\usage{
cache_report(cache_type = NULL, cache_dir = NULL)
}
\arguments{
\item{cache_type}{character. Use \code{temporary} to a session temporary folder, \code{persistent} for a persistent cache folder or \code{NULL} to use the environment default. Check the section Cache Type for more details.}

\item{cache_dir}{character. User specified cache directory for persistent cache type.}
}
\value{
No return value. The function prints a report on the console.
}
\description{
Print a report at the console with cache statistics
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mirror_deposit(deposit_id = 10959197)
cache_report()
\dontshow{\}) # examplesIf}
}
