% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_md5sum.R
\name{check_md5sum}
\alias{check_md5sum}
\title{Check the MD5SUM of a downloaded file against the Zenodo MD5SUM informed value}
\usage{
check_md5sum(file, original_checksum)
}
\arguments{
\item{file}{character. The downloaded file path.}

\item{original_checksum}{character. The MD5SUM informed by Zenodo.}
}
\value{
logical. \code{TRUE} if the MD5SUM matches.
}
\description{
Check the MD5SUM of a downloaded file against the Zenodo MD5SUM informed value
}
