% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_mirror.R
\name{delete_mirror}
\alias{delete_mirror}
\title{Delete a deposit mirror}
\usage{
delete_mirror(deposit_id)
}
\arguments{
\item{deposit_id}{numeric. The Zenodo deposit id.}
}
\value{
No return value. The function deletes mirrored on the system.
}
\description{
This will delete all mirrored files stored locally.
}
