% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_deposit.R
\name{download_deposit}
\alias{download_deposit}
\title{Download files from a Zenodo deposit}
\usage{
download_deposit(list_deposit, file_name = NULL, dest, quiet = FALSE)
}
\arguments{
\item{list_deposit}{data.frame or tibble. Obtained with \link{list_deposit}.}

\item{file_name}{character. If \code{NULL}, all files from the file list. If a file name is specified, only this file will be downloaded.}

\item{dest}{character. Destination folder.}

\item{quiet}{logical. Show download info and progress bar.}
}
\description{
Download files from a Zenodo deposit
}
