% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_list.R
\name{file_list}
\alias{file_list}
\title{Get a list of files available at a Zenodo deposition}
\usage{
file_list(deposit_id)
}
\arguments{
\item{deposit_id}{numeric. The Zenodo deposition id.}
}
\value{
a tibble.
}
\description{
Get a list of files available at a Zenodo deposition
}
