% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_deposit.R
\name{mirror_deposit}
\alias{mirror_deposit}
\title{Mirror a Zenodo deposit locally}
\usage{
mirror_deposit(
  deposit_id,
  file_name = NULL,
  clear_cache = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{deposit_id}{numeric. The Zenodo deposit id.}

\item{file_name}{character. If \code{NULL}, all files from the file list. If a file name is specified, only this file will be downloaded.}

\item{clear_cache}{logical. If the mirror already exists, clear its content.}

\item{quiet}{logical. Show download info and progress bar.}
}
\value{
a string with the mirror path.
}
\description{
Mirror an entire Zenodo deposit or a specific file locally. The mirror is a folder with the deposit files created at the cache folder of the operating system.
}
