% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zendown.R
\name{zendown}
\alias{zendown}
\title{Download files from a Zenodo deposition list of files}
\usage{
zendown(file_list, file_name = NULL, dest, quiet = FALSE)
}
\arguments{
\item{file_list}{data.frame or tibble. Obtained with \code{file_list}.}

\item{file_name}{character. If \code{NULL}, all files from the file list. If a file name is specified, only this file will be downloaded.}

\item{dest}{character. Destination folder.}

\item{quiet}{logical. Show download info and progress bar.}
}
\value{
the downloaded file(s).
}
\description{
Download files from a Zenodo deposition list of files
}
