% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_mirror.R
\name{delete_all_mirrors}
\alias{delete_all_mirrors}
\title{Delete all deposit mirrors}
\usage{
delete_all_mirrors(cache_type = NULL)
}
\value{
No return value. The function deletes all deposits mirrored on the system.
}
\description{
This will delete all deposit mirrors stored locally.
}
\examples{
\dontshow{if (curl::has_internet() & RCurl::url.exists("https://zenodo.org/records/10959197", timeout.ms = 5000)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# https://zenodo.org/records/10959197
file_path <- zen_file(10959197, "iris.rds")
print(file_path)
file.exists(file_path)
delete_all_mirrors()
file.exists(file_path)
\dontshow{\}) # examplesIf}
}
