% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_mirror.R
\name{delete_mirror}
\alias{delete_mirror}
\title{Delete a deposit mirror}
\usage{
delete_mirror(deposit_id)
}
\arguments{
\item{deposit_id}{numeric. The Zenodo deposit id.}
}
\value{
No return value. The function deletes mirrored on the system.
}
\description{
This function will delete all mirrored files stored locally.
}
\examples{
\dontshow{if (curl::has_internet() & RCurl::url.exists("https://zenodo.org/", timeout.ms = 5000)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
zen_file(10959197, "iris.rds")
delete_mirror(10959197)
\dontshow{\}) # examplesIf}
}
