% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen_file.R
\name{zen_file}
\alias{zen_file}
\title{Access a Zenodo deposit file}
\usage{
zen_file(
  deposit_id,
  file_name,
  cache_type = NULL,
  clear_cache = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{deposit_id}{numeric. The Zenodo deposit id.}

\item{file_name}{character. If \code{NULL}, all files from the file list. If a file name is specified, only this file will be downloaded.}

\item{clear_cache}{logical. If the mirror already exists, clear its content.}

\item{quiet}{logical. Show download info and progress bar.}
}
\value{
a string with the mirror file path.
}
\description{
Get the local path to a Zenodo deposit file. It not cached in a mirror yet, it will download the file to a deposit mirror cache locally.
}
\examples{
\dontshow{if (curl::has_internet() & RCurl::url.exists("https://zenodo.org/records/10959197", timeout.ms = 5000)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# https://zenodo.org/records/10959197
zen_file_path <- zen_file(10959197, "iris.rds")
file.exists(zen_file_path)
\dontshow{\}) # examplesIf}
}
